creating a mongo module to isolute this implementation
