+++ pom.xml
@@ -31,6 +31,7 @@
         <module>indra-service</module>
         <module>indra-core</module>
         <module>indra-client</module>
+        <module>indra-mongo</module>
     </modules>
 
     <properties>
